Test get user api
