--- FriendTodo/FriendTodoTests/FTApiReposTest.swift
@@ -1,41 +1,0 @@
-//
-//  FTApiRepoTest.swift
-//  FriendTodoTests
-//
-//  Created by Andrea Bozza on 23/02/2021.
-//
-
-import XCTest
-@testable import FriendTodo
-
-class FTApiRepoTest: XCTestCase {
-
-    override func setUpWithError() throws {
-        
-        
-        func testGetListOfFriendsWithExpectedURLHostAndPath() {
-          let apiRespository = FTApiRepo()
-          apiRepository.getFrieds { friends, error in  }
-        }
-        
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // UI tests must launch the application that they test. Doing this in setup will make sure it happens for each test method.
-        XCUIApplication().launch()
-
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-    }
-
-    func testExample() throws {
-        // Use recording to get started writing UI tests.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-
-}
